Export function for getting GRPC server options
