+++ internal/metamorph/server.go
@@ -8,27 +8,19 @@ import (
 	"log/slog"
 	"net"
 	"os"
-	"runtime/debug"
 	"strings"
 	"sync"
 	"time"
 
+	"github.com/bitcoin-sv/arc/internal/grpc_server"
 	"github.com/bitcoin-sv/arc/internal/metamorph/processor_response"
 	"github.com/bitcoin-sv/arc/internal/metamorph/store"
 	"github.com/bitcoin-sv/arc/pkg/metamorph/metamorph_api"
-	grpcprom "github.com/grpc-ecosystem/go-grpc-middleware/providers/prometheus"
-	"github.com/grpc-ecosystem/go-grpc-middleware/v2/interceptors/logging"
-	"github.com/grpc-ecosystem/go-grpc-middleware/v2/interceptors/recovery"
-	"github.com/grpc-ecosystem/go-grpc-middleware/v2/testing/testpb"
 	"github.com/libsv/go-bt/v2"
 	"github.com/libsv/go-p2p/chaincfg/chainhash"
 	"github.com/ordishs/go-bitcoin"
-	"github.com/prometheus/client_golang/prometheus"
-	"go.opentelemetry.io/otel/trace"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/reflection"
-	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/types/known/emptypb"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
@@ -105,91 +97,18 @@ func (s *Server) SetTimeout(timeout time.Duration) {
 	s.timeout = timeout
 }
 
-// interceptorLogger adapts slog logger to interceptor logger.
-func interceptorLogger(l *slog.Logger) logging.Logger {
-	return logging.LoggerFunc(func(_ context.Context, lvl logging.Level, msg string, fields ...any) {
-		switch lvl {
-		case logging.LevelDebug:
-			l.Debug(msg, fields...)
-		case logging.LevelInfo:
-			l.Info(msg, fields...)
-		case logging.LevelWarn:
-			l.Warn(msg, fields...)
-		case logging.LevelError:
-			l.Error(msg, fields...)
-		default:
-			panic(fmt.Sprintf("unknown level %v", lvl))
-		}
-	})
-}
-
 // StartGRPCServer function
-func (s *Server) StartGRPCServer(address string, grpcMessageSize int, prometheusEndpoint string, loggern *slog.Logger) error {
+func (s *Server) StartGRPCServer(address string, grpcMessageSize int, prometheusEndpoint string, logger *slog.Logger) error {
 	// LEVEL 0 - no security / no encryption
 
-	// Setup logging.
-	rpcLogger := loggern.With(slog.String("service", "gRPC/server"))
-	logTraceID := func(ctx context.Context) logging.Fields {
-		if span := trace.SpanContextFromContext(ctx); span.IsSampled() {
-			return logging.Fields{"traceID", span.TraceID().String()}
-		}
-		return nil
-	}
-
-	// Setup metrics.
-	srvMetrics := grpcprom.NewServerMetrics(
-		grpcprom.WithServerHandlingTimeHistogram(
-			grpcprom.WithHistogramBuckets([]float64{0.001, 0.01, 0.1, 0.3, 0.6, 1, 3, 6, 9, 20, 30, 60, 90, 120}),
-		),
-	)
-
-	exemplarFromContext := func(ctx context.Context) prometheus.Labels {
-		if span := trace.SpanContextFromContext(ctx); span.IsSampled() {
-			return prometheus.Labels{"traceID": span.TraceID().String()}
-		}
-		return nil
-	}
-
-	// Setup metric for panic recoveries.
-	panicsTotal := prometheus.NewCounter(prometheus.CounterOpts{
-		Name: "grpc_req_panics_recovered_total",
-		Help: "Total number of gRPC requests recovered from internal panic.",
-	})
-
-	err := prometheus.Register(panicsTotal)
+	srvMetrics, opts, err := grpc_server.GetGRPCServerOpts(logger, prometheusEndpoint, grpcMessageSize)
 	if err != nil {
-		return fmt.Errorf("failed to register panics total metric: %w", err)
-	}
-
-	grpcPanicRecoveryHandler := func(p any) (err error) {
-		panicsTotal.Inc()
-		rpcLogger.Error("recovered from panic", "panic", p, "stack", debug.Stack())
-		return status.Errorf(codes.Internal, "%s", p)
-	}
-
-	var chainUnaryInterceptors []grpc.UnaryServerInterceptor
-	var chainStreamInterceptors []grpc.StreamServerInterceptor
-
-	if prometheusEndpoint != "" {
-		chainUnaryInterceptors = append(chainUnaryInterceptors, srvMetrics.UnaryServerInterceptor(grpcprom.WithExemplarFromContext(exemplarFromContext)))
-		chainStreamInterceptors = append(chainStreamInterceptors, srvMetrics.StreamServerInterceptor(grpcprom.WithExemplarFromContext(exemplarFromContext)))
-	}
-
-	chainUnaryInterceptors = append(chainUnaryInterceptors, // Order matters e.g. tracing interceptor have to create span first for the later exemplars to work.
-		logging.UnaryServerInterceptor(interceptorLogger(rpcLogger), logging.WithFieldsFromContext(logTraceID)),
-		recovery.UnaryServerInterceptor(recovery.WithRecoveryHandler(grpcPanicRecoveryHandler)))
-	chainStreamInterceptors = append(chainStreamInterceptors, srvMetrics.StreamServerInterceptor(grpcprom.WithExemplarFromContext(exemplarFromContext)),
-		logging.StreamServerInterceptor(interceptorLogger(rpcLogger), logging.WithFieldsFromContext(logTraceID)),
-		recovery.StreamServerInterceptor(recovery.WithRecoveryHandler(grpcPanicRecoveryHandler)))
-	opts := []grpc.ServerOption{
-		grpc.ChainUnaryInterceptor(chainUnaryInterceptors...),
-		grpc.ChainStreamInterceptor(chainStreamInterceptors...),
-		grpc.MaxRecvMsgSize(grpcMessageSize),
+		return err
 	}
 
 	grpcSrv := grpc.NewServer(opts...)
-	t := &testpb.TestPingService{}
-	testpb.RegisterTestServiceServer(grpcSrv, t)
+	//t := &testpb.TestPingService{}
+	//testpb.RegisterTestServiceServer(grpcSrv, t)
 	srvMetrics.InitializeMetrics(grpcSrv)
 
 	s.grpcServer = grpcSrv
@@ -204,12 +123,11 @@ func (s *Server) StartGRPCServer(address string, grpcMessageSize int, prometheus
 	// Register reflection service on gRPC server.
 	reflection.Register(s.grpcServer)
 
-	s.logger.Info("GRPC server listening on", slog.String("address", address))
-
 	go func() {
+		s.logger.Info("GRPC server listening on", slog.String("address", address))
 		err = s.grpcServer.Serve(lis)
 		if err != nil {
-			s.logger.Error("metamorph GRPC server failed", slog.String("err", err.Error()))
+			s.logger.Error("GRPC server failed to serve", slog.String("err", err.Error()))
 		}
 	}()
